Version 1.3.5
31.03.2007 - Process FD could crash if it removed file filters from the
             internal database due to an error in function init_msg_buffer().
             has started up, otherwise starting AFD via init script
             (/etc/init.d/afd) might not start after a system crash.
01.04.2007 - Fix an error in pmatch() and sfilter() function where they
             could not handle the case "r10013*00" r10013301000.
03.04.2007 - If DEFAULT_SMTP_SERVER is set in AFD_CONFIG and none specified
             in DIR_CONFIG, sf_smtp() would fail to send the mail. Fixed.
07.04.2007 - Created new program afd_hex_print to view binary data.
09.04.2007 - Added button to show_olog dialog to view data in archive.
11.04.2007 - Improved set_pw program so it can handle input from mouse
             as well, not just the keyboard.
13.04.2007 - When afdd has no log data to send it sends empty log data
             at regular intervals.
             age.
24.04.2007 - Fix error in 'wait for' directory option.
25.04.2007 - Fix two memory leaks in show_queue.
30.04.2007 - When process return abnormally try pass more information
             what might be the cause of it.
01.05.2007 - udc and uhc now expect a responce from AMG if the action
             was successfull or not.
06.05.2007 - Fix error where afd_mon hangs when reading from fifo.
09.05.2007 - Moved window id from fifodir to users home directory.
16.05.2007 - Calculate correct file size when using 'convert mrz2wmo'.
17.05.2007 - During evaluation of DIR_CONFIG, check if options are correct.
22.05.2007 - Allow user to set some other unit for option archive.
23.05.2007 - Function get_dir_option() returned for 'delete old locked files'
             the value from 'delete queued files'. Fixed.
30.05.2007 - Replaced pmatch function with that from Karl Heuer, which also
             adds support for listing characters in [].
03.06.2007 - Added new status ERROR_ACKNOWLEDGED and ERROR_OFFLINE to
             host status in afd_ctrl dialog.
04.06.2007 - Added fix for HPUX which cannot handle more then one shared
             mmap() from one process to the same file.
05.06.2007 - Use mmap() to write pid's to AFD_ACTIVE file.
10.06.2007 - Added event log.
12.06.2007 - Hopefully improved error handling when file_masks file
             in fifodir gets corrupted.
20.06.2007 - Fix error in initializing socket() function.
21.06.2007 - Split transfer status LED in two halfs indicating direction
             of transfer.
24.06.2007 - Created new dialog to handle events.
26.06.2007 - Added permission check for udc and uhc.
29.06.2007 - When updating DIR_CONFIG it could happen that certain retrieve
             jobs would no longer work until FD was restarted. An additional
             check was added that tries to detect such a situation and
             correct it automaticaly.
01.07.2007 - Do not calculate a new directory scan time if the directory
             was full.
04.07.2007 - Changed meaning of show_dlog and show_elog in afd.users file.
             show_dlog now means the delete log, transfer debug log is
             now set with show_tdlog. show_elog now means event log,
             the delete log is now set with show_dlog.
13.07.2007 - For the extract option it is possible to only extract certain
             files.
15.07.2007 - Implemented new dialog show_elog to view the event log.

+++ src/common/change_name.c
@@ -451,7 +451,7 @@ change_name(char         *orig_file_name,
                            break;
                      }
                      i = 0;
-                     while ((isdigit(*ptr_rule)) && (i < MAX_INT_LENGTH))
+                     while ((isdigit((int)(*ptr_rule))) && (i < MAX_INT_LENGTH))
                      {
                         string[i++] = *ptr_rule++;
                      }
