+++ src/fd/eval_input_sf.c
@@ -93,7 +93,7 @@ eval_input_sf(int argc, char *argv[], struct job *p_db)
    else
    {
       STRNCPY(p_work_dir, argv[1], MAX_PATH_LENGTH);
-      if (isdigit(argv[2][0]) == 0)
+      if (isdigit((int)(argv[2][0])) == 0)
       {
          (void)fprintf(stderr,
                        "ERROR   : None nummeric value for job number : %s.\n",
@@ -111,7 +111,7 @@ eval_input_sf(int argc, char *argv[], struct job *p_db)
          i = 0;
          do
          {
-            if (isdigit(argv[3][i]) == 0)
+            if (isdigit((int)(argv[3][i])) == 0)
             {
                i = MAX_INT_LENGTH;
             }
@@ -125,7 +125,7 @@ eval_input_sf(int argc, char *argv[], struct job *p_db)
             i = 0;
             do
             {
-               if (isdigit(argv[4][i]) == 0)
+               if (isdigit((int)(argv[4][i])) == 0)
                {
                   i = MAX_INT_LENGTH;
                }
@@ -141,7 +141,7 @@ eval_input_sf(int argc, char *argv[], struct job *p_db)
                i = 0;
                do
                {
-                  if ((isxdigit(argv[5][i])) || (argv[5][i] == '_') ||
+                  if ((isxdigit((int)(argv[5][i]))) || (argv[5][i] == '_') ||
                       (argv[5][i] == '/'))
                   {
                      p_db->msg_name[i] = argv[5][i];
@@ -194,7 +194,7 @@ eval_input_sf(int argc, char *argv[], struct job *p_db)
                                        i++;
                                        do
                                        {
-                                          if (isdigit(argv[i][k]) == 0)
+                                          if ((isdigit((int)(argv[i][k]))) == 0)
                                           {
                                              k = MAX_INT_LENGTH;
                                           }
@@ -268,7 +268,7 @@ eval_input_sf(int argc, char *argv[], struct job *p_db)
                                        i++;
                                        do
                                        {
-                                          if (isdigit(argv[i][k]) == 0)
+                                          if (isdigit((int)(argv[i][k])) == 0)
                                           {
                                              k = MAX_INT_LENGTH;
                                           }
